fix: address PR #177 review comments
- Escape document name in exported HTML title and toolbar (XSS)
- Reset perspective to "all" before HTML export to include all relations
- Fix falsy-zero bug in visibleBounds() using !== undefined check
- Wrap HTML export in try/finally to restore canvas transform on error
- Pass drawableArea as prop from MainView instead of exporting from support.ts
- Extract downloadHtml into dedicated htmlExporter.ts module
- Rename data-parent/data-child to data-erd-relation-parent-table-id
  and data-erd-relation-child-table-id
